Add poll_for_run_completion helper
