+++ examples/async_run.py
@@ -1,28 +1,14 @@
-import time
-from datetime import datetime, timedelta
-
 from pipeline.cloud.pipelines import run_pipeline
-from pipeline.cloud.runs import get_run
+from pipeline.cloud.runs import poll_for_run_completion
 from pipeline.cloud.schemas.runs import RunState
 
+pointer = "my_pipeline_id"
+input_data = ["my input string"]
 
-def poll_for_run_completion(run_id, timeout_secs: int = 5 * 60, interval_secs: int = 5):
-    start_time = datetime.now()
-    while datetime.now() - start_time < timedelta(seconds=timeout_secs):
-        run = get_run(run_id)
-        if RunState.is_terminal(run.state):
-            return run
-        print(f"Waiting for run {run_id} to finish...")
-        time.sleep(interval_secs)
-    raise TimeoutError(f"Run {run_id} did not finish in the allotted time")
-
-
-pointer = "pipeline_id"
-
-initial_result = run_pipeline(pointer, "an input string", async_run=True)
+initial_result = run_pipeline(pointer, *input_data, async_run=True)
 run_id = initial_result.id
 
-result = poll_for_run_completion(run_id, interval_secs=1)
+result = poll_for_run_completion(run_id, timeout_secs=5 * 60, interval_secs=10)
 
 if result.state != RunState.completed:
     print(f"Run id: {result.id}, state: {result.state}")
